1、清理一些不用的资源，节省项目流量 2、会话管理持久化，不会内存溢出 3、用户对象缓存处理，加快读取速度 4、解决资源和组织机构删除存在子节点的提醒 5、优化部分界面显示

+++ src/main/java/cn/iutils/sys/controller/OrganizationController.java
@@ -123,10 +123,16 @@ public class OrganizationController extends BaseController {
 	public String delete(Organization organization,int pageNo,int pageSize,
 			RedirectAttributes redirectAttributes) {
 		if(!"1".equals(organization.getId())){
-			organizationService.delete(organization);
-			addMessage(redirectAttributes, "删除成功");
+			//删除前判断是否存在下级节点
+			int count = organizationService.findNext(organization);
+			if(count>0){
+				addMessage(redirectAttributes, "删除失败，存在子节点，请先删除子节点");
+			}else{
+				organizationService.delete(organization);
+				addMessage(redirectAttributes, "删除成功");
+			}
 		}else{
-			addMessage(redirectAttributes, "不能删除");
+			addMessage(redirectAttributes, "超级管理员才能删除");
 		}
 		return "redirect:" + adminPath + "/organization?id="+organization.getParentId()+"&pageNo="+pageNo+"&pageSize="+pageSize;
 	}
